Fix invalid lookup of cardinality attributes. Restore TypeMapper (unused in this commit) Remove BaseRecordReader and specialize between bit and byte aware readers. Restrict key type to (u)int. Correct type of index column in header (I also still do **not** like this approach to headers) Wire in a tiny bit of JetBrain's profiling API.

+++ DBClientFiles.NET/Parsing/File/IHeader.cs
@@ -22,7 +22,7 @@ namespace DBClientFiles.NET.Parsing.File
         int MaxIndex { get; }
 
         int CopyTableLength { get; }
-        int IndexColumn { get; }
+        short IndexColumn { get; }
 
         bool HasIndexTable { get; }
         bool HasForeignIds { get; }
